move filterContacts logic
